fix delphi in case of PHI

+++ DataFormats/Math/interface/deltaPhi.h
@@ -26,7 +26,7 @@ namespace reco {
   
 
   inline float deltaPhi(float phi1, float phi2) { 
-    double result = phi1 - phi2;
+    float result = phi1 - phi2;
     while (result > float(M_PI)) result -= float(2*M_PI);
     while (result <= -float(M_PI)) result += float(2*M_PI);
     return result;
